refactor(d3): relocate parseVisualization + types out of d3.ts (109.2 step 1)
Move the parse state machine to src/visualizations/parse.ts and the parsed
shape types to src/visualizations/types.ts. dgmo-router + chart-type-registry
now import parseVisualization from visualizations/parse, not d3.ts.

This clears BOTH d3<->dgmo-router circular deps (the one Story 109.1 added AND
the pre-existing one) — madge 7 -> 5; check:circular guard tightened 7 -> 5.
d3.ts shrinks 8863 -> ~7240 lines (renderers + export dispatch remain, split
next). Public API unchanged via d3.ts re-export barrel. Zero snapshot drift,
6831 tests pass.
